fixed event order by adding modified to folder and adapting the compare service also fixed the recreatin when deleting bug and renabled the delteIT

+++ pieShareApp/src/main/java/org/pieShare/pieShareApp/task/eventTasks/conflictTasks/FileListTask.java
@@ -15,6 +15,7 @@ import org.pieShare.pieShareApp.model.pieFilder.PieFolder;
 import org.pieShare.pieShareApp.service.fileService.api.IFileService;
 import org.pieShare.pieShareApp.service.folderService.FolderServiceException;
 import org.pieShare.pieShareApp.service.folderService.IFolderService;
+import org.pieShare.pieShareApp.service.historyService.IHistoryService;
 import org.pieShare.pieTools.pieUtilities.service.pieLogger.PieLogger;
 
 /**
@@ -25,6 +26,7 @@ public class FileListTask extends ARequestTask<IFileListMessage> {
 
 	private IFileService fileService;
 	private IFolderService folderService;
+//	private IHistoryService historyService;
 
 	public void setFileService(IFileService fileService) {
 		this.fileService = fileService;
@@ -36,30 +38,26 @@ public class FileListTask extends ARequestTask<IFileListMessage> {
 
 	@Override
 	public void run() {
-
-		//todo-after-ase: this is not entirely correct and only a temporary solution
-			//we need to perform all relevant local checks to ensure that a
-			//folder/file is allowed to be delete otherwise race conditions are possible
-			//+ a folder needs a last modified so that when we know which event has priority
+		//todo-after-ase: this is just a quickfix to make the list sync work at least for delete and create
+		//we will need to implement a full history to ensure proper resync after being offline
 		for (PieFile file : this.msg.getFileList()) {
-			if (file.isDeleted()) {
-				//todo: can not simply delete file!
-					//we need to ask for the history so that we can determin
-					//if we have a conflict or not
-				fileService.deleteRecursive(file);
-			} else {
-				//in future consider starting an own task for requesting files because we don't 
-				//want to block by user resolution all autoresolvable requests
-				this.doWork(file);
+			if (!this.isConflictedOrNotNeeded(file)) {
+				if (file.isDeleted()) {
+					fileService.deleteRecursive(file);
+				} else {
+					//in future consider starting an own task for requesting files because we don't 
+					//want to block by user resolution all autoresolvable requests
+					this.requestService.requestFile(file);
+				}
 			}
 		}
 
 		for (PieFolder folder : this.msg.getFolderList()) {
 			try {
-				if (folder.isDeleted()) {
-					this.folderService.deleteFolder(folder);
-				} else {
-					if(!this.folderService.getAbsolutePath(folder).exists()) {
+				if (!this.comparerService.isConflictedOrNotNeeded(folder)) {
+					if (folder.isDeleted()) {
+						this.folderService.deleteFolder(folder);
+					} else if (!this.folderService.getAbsolutePath(folder).exists()) {
 						this.folderService.createFolder(folder);
 					}
 				}
@@ -68,4 +66,23 @@ public class FileListTask extends ARequestTask<IFileListMessage> {
 			}
 		}
 	}
+
+	//todo-after-ase: this is also part of the quick fix! consider this two options:
+	//1. extend compare service to also have functionality for folders or
+	//2. remove compare service entirely and add the logic where needed in place like here
+	//this would be possible due to the fact that the compare service only covers the retrieval of the
+	//local or history object and performs otherwise the same as
+//	private boolean conflictedOrNotNeeded(PieFolder folder) {
+//		PieFolder localFolder = this.historyService.getPieFolder(folder.getRelativePath());
+//
+//		if (localFolder == null) {
+//			return false;
+//		}
+//
+//		if (localFolder.compareTo(folder) == -1) {
+//			return false;
+//		}
+//
+//		return true;
+//	}
 }
